+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- In `arange`, the array type can be deduced from its arguments if no bit-
+  specifiers are given.
+
 ### Fixed
 
 ### Changed